refactor: LineSections / 구간들이 특정 노선에 속하는지 확인한다.
