+++ README.md
@@ -39,7 +39,6 @@
 - [x] 상행 종점역과 하행 종점역만 있는 경우(구간이 1개인 경우)인데 삭제 시 예외 발생시킨다
 - [x] 삭제할 역이 노선의 하행 종점역이 아니면 삭제할 수 없다
 - [x] 끊어진 구간들로 생성할 수 없다
-- [x] 다른 노선에 속한 구간들로 생성할 수 없다
 
 ### 지하철 노선(line)
 지하철 구간의 모음으로 구간에 포함된 지하철 역의 연결 정보
@@ -48,7 +47,7 @@
 지하철 노선 구간을 관리한다.
 
 - [x] 노선을 생성할 때 한 구간을 생성한다.
-- [x] 현재 구간은 해당 노선에 속하지 않습니다.
+- [x] 구간들이 특정 노선에 속하는지 확인한다.
 
 ### 체크 리스트
 - Sections 빈 리스트 예외 처리 확인
