+++ src/test/java/subway/domain/LineSectionsTest.java
@@ -1,5 +1,8 @@
 package subway.domain;
 
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -30,5 +33,16 @@ class LineSectionsTest {
 
         Assertions.assertThatCode(() -> new LineSections(lineA, section))
             .doesNotThrowAnyException();
+    }
+
+    @Test
+    @DisplayName("다른 노선에 속한 구간들로 생성할 수 없습니다.")
+    void cannotCreateWithSectionsOfOtherLines() {
+        Line otherLine = new Line(2L, "lineB", "red");
+        Section sectionA = new Section(lineA, stationA, stationB, 3);
+        Section sectionB = new Section(otherLine, stationB, stationC, 3);
+
+        assertThatThrownBy(() -> new LineSections(lineA, new Sections(List.of(sectionA, sectionB))))
+            .isInstanceOf(IllegalArgumentException.class);
     }
 }
